Fix: Refactor VMPool class to decouple it from orchestrator path.
