+++ src/aleph/vm/orchestrator/run.py
@@ -18,9 +18,9 @@ from aleph.vm.hypervisors.firecracker.microvm import MicroVMFailedInit
 from aleph.vm.utils import HostNotFoundError
 
 from ..conf import settings
+from ..pool import VmPool
 from .messages import load_updated_message
 from .models import VmExecution
-from .pool import VmPool
 from .pubsub import PubSub
 
 logger = logging.getLogger(__name__)
